requirements.txt
Faker==39.0.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
tzdata==2025.3

requests
pyyaml
boto3
kafka-python
fastparquet
pandas
redis

dbt-snowflake
